5186 Extract InvestigateCompactionChangingNumberOfRecords
